[finishes 165020212] resolves the conficts and apply the review changes
