+++ tests/dataGenerator.js
@@ -1,5 +1,9 @@
 // import uuid from 'uuid';
 import crypto from 'crypto';
+import moment from 'moment';
+
+const createdAt = moment('2018-01-07').format();
+const updatedAt = createdAt;
 
 const salt = crypto.randomBytes(16).toString('hex');
 const password = 'testuser';
@@ -41,90 +45,119 @@ export default {
   },
   // post1 published
   post1: {
-    id: 'f36e2ee8-e206-4ccb-a6ec-dd752ea45ec3',
-    title: 'the complete junior to senior web',
+    id: 'c90dee64-663d-4d8b-b34d-12acba22cd30',
+    author: 'dfef16f9-11a7-4eae-9ba0-7038c6ccaa73',
+    title: 'this is my first try of article',
+    slug: 'this-is-my-first-try-of-article69f9fccd65',
+    description: 'new article',
     body:
-      'In order to keep up with the industry, your best bet is to be efficient and be wise about what you spend time on because it is impossible to learn and know everything.',
-    author: 'acb2f760-aa40-4983-a664-31c520eb656f',
-    slug: 'the-complete-junior-81234',
-    description: 'In order to keep up with the industry, your best bet is to be efficient',
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla faucibus ipsum non metus finibus ultricies. Donec ac auctor dui, sed fringilla est. Duis et pellentesque nisl, a gravida felis. Ut tempor felis id dignissim congue. Nunc blandit nunc sit amet dui pharetra, quis porttitor sem ullamcorper. Suspendisse faucibus imperdiet lacinia.',
     status: 'published',
+    tagList: JSON.stringify(['lorem']),
     deleted: 0,
-    tagList: JSON.stringify(['developer'])
+    createdAt,
+    updatedAt
   },
   // post2 published
   post2: {
-    id: '0bdbf765-3dcb-4d39-8fd6-f268d9fdebab',
-    title: 'Lorem ipsum dolor sit',
+    id: 'c90dee64-663d-4d8b-b34d-12acba22cd31',
+    author: 'dfef16f9-11a7-4eae-9ba0-7038c6ccaa73',
+    title: 'this is my second try of article',
+    slug: 'this-is-my-second-try-of-article69f9fccd65',
+    description: 'second article',
     body:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua.',
-    author: 'acb2f760-aa40-4983-a664-31c520eb656f',
-    slug: 'lorem-ipsum-dolor-sit-45033',
-    description:
-      'Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt',
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla faucibus ipsum non metus finibus ultricies. Donec ac auctor dui, sed fringilla est. Duis et pellentesque nisl, a gravida felis. Ut tempor felis id dignissim congue. Nunc blandit nunc sit amet dui pharetra, quis porttitor sem ullamcorper. Suspendisse faucibus imperdiet lacinia.',
     status: 'published',
+    tagList: JSON.stringify(['lorem']),
     deleted: 0,
-    tagList: JSON.stringify(['lorem'])
+    createdAt,
+    updatedAt
   },
   // post3 drafted
   post3: {
-    id: '6e7a2d6c-4118-4f17-868e-8669320466df',
-    title: 'Duis aute irure dolor',
+    id: 'c90dee64-663d-4d8b-b34d-12acba22cd32',
+    author: 'dfef16f9-11a7-4eae-9ba0-7038c6ccaa73',
+    title: 'this is my first disliked article',
+    slug: 'this-is-my-first-disliked-article69f9fccd65',
+    description: 'first disliked article',
     body:
-      'Duis aute irure dolor in reprehenderit in voluptate velit esse cillum dolore eu fugiat nulla pariatur.',
-    author: 'acb2f760-aa40-4983-a664-31c520eb656f',
-    slug: 'duis-aute-irure-dolor-fre453972',
-    description: 'Duis aute irure dolor in reprehenderit',
-    deleted: 0,
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla faucibus ipsum non metus finibus ultricies. Donec ac auctor dui, sed fringilla est. Duis et pellentesque nisl, a gravida felis. Ut tempor felis id dignissim congue. Nunc blandit nunc sit amet dui pharetra, quis porttitor sem ullamcorper. Suspendisse faucibus imperdiet lacinia.',
     status: 'published',
-    tagList: JSON.stringify([])
+    tagList: JSON.stringify(['lorem']),
+    deleted: 0,
+    createdAt,
+    updatedAt
   },
   // post4 to update
   post4: {
-    id: 'fd3bc6a5-74d4-4121-92be-72f7829b62b3',
-    title: 'the complete junior to senior web guide',
+    id: 'c90dee64-663d-4d8b-b34d-12acba22cd33',
+    author: 'dfef16f9-11a7-4eae-9ba0-7038c6ccaa73',
+    title: 'this is my second disliked article',
+    slug: 'this-is-my-second-disliked-article69f9fccd65',
+    description: 'second disliked article',
     body:
-      'In order to keep up with the industry, your best bet is to be efficient and be wise about what you spend time on because it is impossible to learn and know everything.',
-    author: 'acb2f760-aa40-4983-a664-31c520eb656f',
-    slug: 'the-complete-junior-guide-81204',
-    description: 'In order to keep up with the industry, your best bet is to be efficient',
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla faucibus ipsum non metus finibus ultricies. Donec ac auctor dui, sed fringilla est. Duis et pellentesque nisl, a gravida felis. Ut tempor felis id dignissim congue. Nunc blandit nunc sit amet dui pharetra, quis porttitor sem ullamcorper. Suspendisse faucibus imperdiet lacinia.',
     status: 'published',
+    tagList: JSON.stringify(['lorem']),
     deleted: 0,
-    tagList: JSON.stringify(['developer'])
+    createdAt,
+    updatedAt
   },
   // post5 to delete
   post5: {
-    id: '622119ea-f74b-440c-b633-917ac5946dc9',
-    title: 'Excepteur sint occaecat cupidatat',
+    id: 'c90dee64-663d-4d8b-b34d-12acba22cd34',
+    author: 'dfef16f9-11a7-4eae-9ba0-7038c6ccaa73',
+    title: 'this is my first favorite article',
+    slug: 'this-is-my-first-favorite-article69f9fccd65',
+    description: 'second disliked article',
     body:
-      'Excepteur sint occaecat cupidatat non proident, sunt in culpa qui officia deserunt mollit anim id est laborum',
-    author: 'acb2f760-aa40-4983-a664-31c520eb656f',
-    slug: 'excepteur-sint-occaecat-cupidat-356d344',
-    description: 'Excepteur sint occaecat cupidatat non proident',
-    deleted: 0,
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla faucibus ipsum non metus finibus ultricies. Donec ac auctor dui, sed fringilla est. Duis et pellentesque nisl, a gravida felis. Ut tempor felis id dignissim congue. Nunc blandit nunc sit amet dui pharetra, quis porttitor sem ullamcorper. Suspendisse faucibus imperdiet lacinia.',
     status: 'published',
-    tagList: JSON.stringify(['uncommon'])
+    tagList: JSON.stringify(['lorem']),
+    deleted: 0,
+    createdAt,
+    updatedAt
+  },
+  post6: {
+    id: 'c90dee64-663d-4d8b-b34d-12acba22cd35',
+    author: 'dfef16f9-11a7-4eae-9ba0-7038c6ccaa73',
+    title: 'this is my second favorite article',
+    slug: 'this-is-my-second-favorite-article69f9fccd65',
+    description: 'second disliked article',
+    body:
+      'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla faucibus ipsum non metus finibus ultricies. Donec ac auctor dui, sed fringilla est. Duis et pellentesque nisl, a gravida felis. Ut tempor felis id dignissim congue. Nunc blandit nunc sit amet dui pharetra, quis porttitor sem ullamcorper. Suspendisse faucibus imperdiet lacinia.',
+    status: 'published',
+    tagList: JSON.stringify(['lorem']),
+    deleted: 0,
+    createdAt,
+    updatedAt
   },
   // comment1 to published story
   comment1: {
     id: '3797fca1-95fc-49fe-af38-787e880ca1d3',
     body: 'Thanks fo the article, its a great article',
-    author: 'acb2f760-aa40-4983-a664-31c520eb656f',
-    post: 'f36e2ee8-e206-4ccb-a6ec-dd752ea45ec3'
+    author: 'dfef16f9-11a7-4eae-9ba0-7038c6ccaa73',
+    post: 'c90dee64-663d-4d8b-b34d-12acba22cd30',
+    createdAt,
+    updatedAt
   },
   // comment2 to published story
   comment2: {
     id: '18e1bbf9-e707-4925-a992-c59f1fc748aa',
     body: 'Hey tester, I want to know how you do tests?',
-    author: 'acb2f760-aa40-4983-a664-31c520eb656f',
-    post: 'f36e2ee8-e206-4ccb-a6ec-dd752ea45ec3'
+    author: 'dfef16f9-11a7-4eae-9ba0-7038c6ccaa73',
+    post: 'c90dee64-663d-4d8b-b34d-12acba22cd30',
+    createdAt,
+    updatedAt
   },
 
   // comment3 to delete
   comment3: {
     id: '2f0bab3d-54f0-41bb-b20e-b3456b28342f',
     body: 'It can take another post, I will work on it',
-    author: 'acb2f760-aa40-4983-a664-31c520eb656f',
-    post: 'f36e2ee8-e206-4ccb-a6ec-dd752ea45ec3'
+    author: 'dfef16f9-11a7-4eae-9ba0-7038c6ccaa73',
+    post: 'c90dee64-663d-4d8b-b34d-12acba22cd30',
+    createdAt,
+    updatedAt
   }
 };
